Replaced old models and moved transforms
